Refactor project routing and screens: Rename ProjectsScreen to ProjectScreen, update routing constants, and create a new PaymentPlansScreen.
