+++ lib/core/routes/routes.dart
@@ -11,5 +11,5 @@ class RoutesNames {
   static const String contacts = '/contacts';
   static const String campaigns = '/campaigns';
   static const String tasks = '/tasks';
-  static const String projects = '/projects';
+  static const String project = '/project';
 }
