+++ lib/core/widgets/app_menu.dart
@@ -120,7 +120,7 @@ class AppMenu extends StatelessWidget {
       'Clients': RoutesNames.clients,
       'Campaigns': RoutesNames.campaigns,
       'Tasks': RoutesNames.tasks,
-      'Projects': RoutesNames.projects,
+      'Projects': RoutesNames.project,
       //? more routes here
     };
     for (MapEntry menus in R.icons.iconsMenu.entries) {
